+++ src/hooks/useToggle.js
@@ -1,10 +1,10 @@
 import React from 'react';
 
-function useBoolValue(initialValue) {
+function useToggle(initialValue) {
     const [value, setValue] = React.useState(initialValue);
     const toggleValue = () => setValue(!value);
 
     return [value, toggleValue];
 }
 
-export default useBoolValue;
+export default useToggle;