+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## [0.0.6] - 20/04/2021
+
+* Renamed `useBoolValue` to `useToggle`
+
+## [0.0.5] - 20/04/2021
+
+* Consolidated code standards across hooks
+
+## [0.0.4] - 20/04/2021
+
+* Fixed hook exporting issue
+
 ## [0.0.3] - 20/04/2021
 
 * Added `constants.js` file to avoid magic string implementation